'Direct jars' should be a subset of recursively collected jars
java_common.create_provider currently does not add the provided direct jars to the
set of transitively collected jars, which is inconsistent with how the two sets are
used in the native rule implementations. The direct jars should be the subset of
transitive jars that corresponds to direct deps, they are not supposed to
be disjoint.

RELNOTES: in the JavaInfo created by java_common.create_provider now includes both direct and transitive arguments in transitive_compile_time_jars and transitive_runtime_jars
PiperOrigin-RevId: 195123581