Harden Profiler slow tasks capping test
Internally slow task aggregators are sharded by thread id, so we need to throw
some threads at it to get an interesting test case, otherwise everything comes
from one thread and lands in one shard and the shard combination logic isn't
exercised. Of course it's possible that all the threads created in the test
wind up landing in the same shard, let's just pretend that's not going to
happen, since the odds are pretty low.

PiperOrigin-RevId: 195141874