CROSSTOOLS: wrapped_clang remove CLI parsing
Before commit 53700e240a Bazel was quoting some calls to wrapped_clang
like so:

  '-isysroot /some/path maybe with spaces'

and wrapped_clang was parsing to split on spaces (assuming the paths themselves
didn't have spaces). After 53700e240a the arguments to wrapped_clang don't
need to be parsed so this commit removes the parsing.

Also add a test for spaces in pathnames.