Some fixup for R.class files
Expose R.class file in AndroidResourcesInfo, to allow Java compilation.

Include documentation that the R.class should not be used in other targets.

Actually generate the R.class file in the case of an android_library with no
manifest, resources, or assets, rather than compiling it from the R.java

RELNOTES: none
PiperOrigin-RevId: 195419781