Properly expose R.class file to Skylark
Add the R.class to the provider, but keep it nullable in case no
resource/manifest attributes were specified, in which case the R.class would
not be generated.

A previous change tried to actually generate it in that case, but even though I
think that's probably the right strategy for the long term, it's substantially
different behavior and out of scope of this migration.

RELNOTES: none
PiperOrigin-RevId: 195682271