Expose R.class as a JavaInfo
The R.class file, when available, is used in Java compilation. So it should be
exposed in its own provider as an output of resource processing.

The R.class file is marked as Neverlink and compiletime-only to ensure it
doesn't get inherited.

Change resource processing to return a SkylarkDict of providers. That way,
callers can easily get the provider they want, or use values() to get all the
providers.

Stop exposing R.class in AndroidResourcesInfo since we're exposing it in an
easier-to-consume way now.

Will update other methods that return multiple providers to do the same in
future reviews.

RELNOTES: none
PiperOrigin-RevId: 195871587