Properly rename manifests
Make AndroidSkylarkData abstract and add a getAndroidSemantics() method

The Skylark stamp_manifest rule should properly use AndroidSemantics to rename
the manifest as well, in keeping with the android_library rule.

AndroidLocalTest should do the same manifest renaming regardless of which
AndroidSemantics it is run under. Have it use a seperate method to do so.

RELNOTES: none
PiperOrigin-RevId: 195992300