Implement a limit for named_set_of_file events
This adds a command-line option which can be used to force Bazel to split up
very large events, e.g., events with hundreds of thousands of entries. This is
rare, but happens occasionally.

It would be better to control the maximum event size directly, but that is
significantly more difficult to do here.

PiperOrigin-RevId: 196690600