Fork RuleContext.ErrorReporter to create a Skylark-friendly RuleErrorConsumer
The new error consumer is simple to create from @SkylarkCallable methods, so we
won't need to pass SkylarkRuleContext into those methods for that reason
anymore.

Create a new abstract class, rather than expand the existing
AbstractRuleErrorConsumer class, since the latter is already being extended and
used in a variety of tests, and I prefer not to have to migrate all of them as
needed. Add Javadoc to make clear that this class should not be used similarly.

RELNOTES: none
PiperOrigin-RevId: 197148849