cpp runfiles library: precompute envvars map
Compute the list of environment variables when
creating the runfiles library instead of computing
it on-demand.

Since we already have the manifest and directory
path from Runfiles::PathsFrom, creating the vector
is cheap, though we have to store it even if we
don't use it later.

In a subsequent change I'll merge the manifest-
and directory-based implementations into
RunfilesImpl, and collapse RunfilesImpl into the
Runfiles base class. The point of that is to make
the runfiles library be able to handle both the
directory-based and the manifest-based case
simultaneously, dependening on what is available
to it in the filesystem.

See https://github.com/bazelbuild/bazel/issues/4460

Change-Id: I56310423528df2d0f7494f45904150193368e2d6

Closes #5218.

Change-Id: I73d1f44611c4b0a73c41162319d0ba7a8a4ae6d7
PiperOrigin-RevId: 197543865