Create AndroidDataContext to wrap context needed when working with Android data
This new object wraps Label, ActionConstructionContext, and various
data-related tools (currently, just the resource Busybox and the Android SDK).
All three of these things are required for creating pretty much any action for
working with Android assets, resources, and manifests, so it makes sense to
pass them around together.

Eventually, we'll use this class instead of RuleContext in any code called from
the Skylark APIs. For now, though, just have it wrap a RuleContext - the
migration away from that will happen in future changes.

Change the current Skylark API to take this class, rather than a
SkylarkRuleContext. This is the first step towards not passing RuleContext (or
SkylarkRuleContext) into the Skylark API.

RELNOTES: none
PiperOrigin-RevId: 197549155