Pass AndroidDataContext through new asset and resource pipelines
This is the first step towards using AndroidDataContext in all of Android data
processing.

This change does not actually modify the asset and resource processing action
builders themselves - they will be migrated in an upcoming change.

Also, add AndroidSemantics to some rules so they can make an
AndroidDataContext.

RELNOTES: none
PiperOrigin-RevId: 197555938