runfiles,cc: Runfiles uses manifest AND directory
Change the C++ runfiles library so it can use the
runfiles manifest (if present) and the runfiles
directory (if present) simultaneously.

If the Runfiles object fails to look up a runfile
from the manifest (or the manifest didn't exist or
wasn't found) then it looks it up from the
runfiles directory.

This change allows using the same binary with and
without a runfiles tree, which is often what sets
local and remote runs apart.

https://github.com/bazelbuild/bazel/issues/4460

Change-Id: Iae879ff084ba084fcd7c111638ddeae4c6754f4f