Split user_link_flags from legacy_link_flags
The difference between them is that user_link_flags will stay after we remove
legacy fields from the crosstool.

This is an encore of https://github.com/bazelbuild/bazel/commit/2661abb96b1fe51fb726a63eb08698564a82eb20 after mitigating the memory regression.

Original cl regressed 12mb (on a big enough internal project :):

 objsize  chg   instances       space KB    class name
------------------------------------------------------
      24   +0      190265        4459 KB    com.google.common.collect.ImmutableMapEntry
      40   +0       95154        3716 KB    com.google.common.collect.RegularImmutableMap
      47   -2       95154        2230 KB    [Ljava.util.Map$Entry;
      44   -1       95154        2230 KB    [Lcom.google.common.collect.ImmutableMapEntry;
      24   +0       95149        2230 KB    com.google.devtools.build.lib.rules.cpp.CcToolchainVariables$MapVariables
      16   +0       95149        1486 KB    com.google.devtools.build.lib.rules.cpp.CcToolchainVariables$StringSequence
      89   +0        3188         176 KB    [B
      64   -8           0        -743 KB    com.google.devtools.build.lib.rules.cpp.LinkCommandLine
      76   +0       -1918       -1323 KB    [Ljava.lang.Object;
      24   +0      -95149       -2230 KB    com.google.devtools.build.lib.rules.cpp.CcToolchainVariables$SingleVariables
------------------------------------------------------

The reason was that before legacy_link_flags were the only build variable
patched because of thinlto in CppLinkActionBuilder, so it used optimized
SingleVariables subclass.  With the split, the patched variables instance had 2
variables therefore it received the full blown Variables instance. That
accounted for the ~7mb.

The fix was to move the patching logic to the initial link variables creation
and not to create patched variables at all. Now the regression is ~4.8mb, which
is perfectly expected since I introduce another variable and this is the
overhead of additional hashmap entry:

 objsize  chg   instances       space KB    class name
------------------------------------------------------
      24   +0      190418        4462 KB    com.google.common.collect.ImmutableMapEntry
      44   +1          31        1487 KB    [Lcom.google.common.collect.ImmutableMapEntry;
      16   +0       95149        1486 KB    com.google.devtools.build.lib.rules.cpp.CcToolchainVariables$StringSequence
      47   +0          31         744 KB    [Ljava.util.Map$Entry;
      89   +0        5723         315 KB    [B
      24   +0        5721         134 KB    java.lang.String
      64   -8           0        -743 KB    com.google.devtools.build.lib.rules.cpp.LinkCommandLine
      76   +0       -2387        -840 KB    [Ljava.lang.Object;
      24   +0      -95149       -2230 KB    com.google.devtools.build.lib.rules.cpp.CcToolchainVariables$SingleVariables
------------------------------------------------------

I'll pay this dept back once we get rid of legacy crosstool fields (= removal of
legacy_link_flags variable).

RELNOTES: None.
PiperOrigin-RevId: 197574153