Skylark: do not eagerly scan the whole file
With this change, the parser explicitly asks the lexer to give the next token.
To avoid changing the lexer too much, the tokenize() method populates a queue (it may add multiple tokens at the same time). While this reduces the peak memory usage, further work is needed to actually improve the performance.

RELNOTES: None.
PiperOrigin-RevId: 197576326