[Skylark] Size newly created arrays to avoid unnecessary re-allocations.
It's more efficient to use properly sized array builders to avoid having
a need to dynamically adjust the size of the underlying array. The result
is improved performance (no need to copy elements to newly created array)
and better memory efficiency - no need to re-allocate and no extra space
is wasted.