For compilations with C++ modules, treat private headers (header files in 'srcs') as textually included headers.
A private header is never exported to a different module and thus cannot be
used by a different module. We never need to compile a module just because one
of it's private headers is used from somewhere else. There must be a different
cc_library (and thus module) that exports the header.

At the same time, being compiled into a module doesn't provide benefits for the
compilation of that cc_library itself, because a module created for a
cc_library A is never used when compiling any of A's srcs.

PiperOrigin-RevId: 197846388