shallow_since tests and doc improvements
This changeset introduces tests for the `shallow_since` functionality that was introduced in 6496b1d1d25025f33406b1eaf9949cab126f08bd.

We test that if the user specifies a tag as well as a `shallow_since` that it fails since they are incompatible. We also verify that specifying a commit before the `shallow_since` date causes the clone to fail (albeit currently with a non-useful error message; perhaps that could be improved at a later date).

It also documents more clearly how to use the feature by explaining the format of the 'since' string.

Closes #5183.

PiperOrigin-RevId: 197875727