Reject files when the first line is indented.
A bug in the lexer ignored indentation on the first line of a file. This now
causes an error.

Also, remove the COMMENT token from the lexer. Comments are now accessed separately. This will allow further optimizations in the lexer. It also aligns the code a bit more with the Go implementation.

RELNOTES[INC]:
  Indentation on the first line of a file was previously ignored. This is now fixed.

PiperOrigin-RevId: 197889775