Make @SkylarkModule detection (through superclasses/superinterfaces) well-defined.
Moving forward, if a class not already annotated with @SkylarkModule has SkylarkModule supertypes A and B, then it must be the case that A is a type of B, or B is a type of A. The skylark type of a given class is dictated by the *most specific* superclass annotated with @SkylarkModule.

RELNOTES: None.
PiperOrigin-RevId: 197946898