Fix build and run of RealSandboxfsProcessTest.
https://github.com/bazelbuild/bazel/commit/56d1b1c3122a3d1ec111baab339631b8c42c2c31 changed the signature of SandboxfsProcess#mount() to take a
PathFragment instead of a Path, and this broke the build of the test above.
Given that the test is marked as manual, this was not caught by CI.

Further, the test was failing to link in the test runner class, so it
would not execute properly.  The missing dependency was incorrectly removed
by unknown commit because it had been specified as a deps instead of a
runtime_deps.

RELNOTES: None.
PiperOrigin-RevId: 198048748