Remove java support from proto_library.
This is a rollforward of CL/197725926 after depot fix.

NEW: Additional deletions of unused providers (no constructor call references).

NEW[last rollforward]: Allow java_* rules to depend on proto_libraries via runtime_deps and exports. This should avoid the breakage that caused the original rollback. The edges are no-ops and could be removed.

PiperOrigin-RevId: 198060345