Enable --experimental_remote_spawn_cache by default
- It is now an error to specify the gRPC remote execution backend in
  combination with a local disk or HTTP-based cache.
- It is now an error to specify both local disk and HTTP-based caches.

Note that before this CL, enabling the local disk cache silently disabled
remote execution - we now give an error in that case.

With these combination no longer being accepted, remote execution being enabled
now means that we only create a RemoteSpawnRunner, and don't provide a
SpawnCache. This is not a semantic change - we never created both.

In principle, it should be possible for users to combine local execution with
remote caching for actions that are marked local or no-remote, and still use
remote execution otherwise. However, Bazel cannot currently express this
combination of execution strategies.

RELNOTES: The --experimental_remote_spawn_cache flag is now enabled by default, and remote caching no longer needs --*_strategy=remote flags (it will fail if they are specified).
PiperOrigin-RevId: 198280398