Add clang headermap support
This is a squashed version of PR #3712
https://github.com/bazelbuild/bazel/pull/3712
by Jerry Marino <i@jerrymarino.com>

It adds an objc_library option called flatten_virtual_headers
that modifies the behavior of include_prefix such that instead
of prepending the path to the header's actual path, the headers
are mapped to the include_path namespace. E.g.

  src/heaaders/foo/bar.h src/headers/bar/baz.h

with the options:

  include_prefix = "FOO"
  flatten_virtual_headers = True

would let you include bar.h and baz.h as

  #include <FOO/bar.h>
  #include <FOO/baz.h>