[java_common.compile] Always generate a source jar.
java_common.compile doesn't generate the output source jar when a source jar is
the only input for the compilation. This is wrong because the source jar can
include APT generated sources. It is also inconsistent with java_library and
leads to inconsistent Skylark rules where a declared output will not always have a
generating action.

RELNOTES: None.
PiperOrigin-RevId: 198872718