Move --distdir out of experimental status
The option --experimental_distdir has been introduced 4 months
ago and was completely unproblematic ever since. Moreover, it
is now used productively, both in our own bootstrapping process[1],
as well as in external packaging of projects using bazel[2]. So
make this option non-experimental. We still keep the old name as
an alternative to not break existing uses.

Related: #5175.

RELNOTES: The --distdir option is no longer experimental. This
  option allows to specify additional directories to look for
  files before trying to fetch them from the network. Files from
  any of the distdirs are only used if a checksum for the file
  is specified and both, the filename and the checksum, match.

[1] Commit 3c9cd82b847f3ece8ec04b2029bd5e8ad0eb7502
[2] https://github.com/gentoo/gentoo/blob/7379cdb578b0c070c846c3fa9f71470e2c5d1320/sci-libs/tensorflow/tensorflow-1.8.0-r1.ebuild#L168

Change-Id: I536238f9bdbad6b4f7222b4f6a1464d70d9f3be3
PiperOrigin-RevId: 199637265