Inline some ActionContextProvider classes into their modules
This also gets rid of some boilerplate. The ExecutionTool.addActionContext
method has been around for a while, but is underused.

There are still a few ActionContextProvider implementations left, which are
implementing other functionality besides adding action contexts.

As a side effect, this change reduces null build time with a hot server on
linux by about a quarter. We were running the linux sandbox twice on every
build, which takes about 70ms each (on my machine), with the total null build
time around 300ms.

PiperOrigin-RevId: 200045145