Implement LcovMerger.
LcovMerger is a tool that merges all the intermediate lcov tracefiles (with .dat extension) found under a coverage directory and prints the merged tracefile to a given output file.

A custom implementation for merging lcov tracefiles is needed because the merging functionality of lcov itself is very slow.

LcovMerger is required to get a single coverage report (lcov tracefile) from a bazel coverage command that executes multiple tests.

ATM LcovMerger is only invoked by tools/test/collect_coverage.sh that collects and merges the tracefiles from a single test invocation. It will also be used from a CoverageReportAction.

Progress on #5246.

PiperOrigin-RevId: 200054506