Windows: fix "corrupt installation" at new year
Bazel on Windows is now consistent with Bazel on
Unixes, by setting the mtimes of embedded binaries
to 10 years in the future, instead of setting them
to January 1st in the 10th year from now.

On Unix, Bazel sets the mtime of embedded binaries
to 10 years into the future. At startup, it checks
that their mtime is at least 9 years in the
future, for two reasons:

- to detect unintentional tampering with the
  embedded binaries, which would reset their mtime
  to the current time, and

- and because we don't expect users to use the
  same version of Bazel in the same workspace for
  more than a year.

On Windows, Bazel had a bug: instead of setting
the mtime to the current time + 10 years, and
comparing them to current time + 9 years, it was
setting current year + 10 years, January 1st, and
comparing to current year + 9 years, January 1st.

This meant that if a user ran Bazel for the first
time on 2017/12/29, then ran again on 2018/01/02,
then:

- on Unixes, the mtimes were set to 2027/12/29,
  then later compared to 2027/01/02, but

- on Windows, the mtimes were set to 2027/01/01,
  then later compared to 2027/01/01, of which
  they were not newer.

Fixes https://github.com/bazelbuild/bazel/issues/4378

Change-Id: I3457bdc360a62a279d1d08c9a69997929f2067dd